refactor: consolidate loading spinner
