+++ app/legacy/onboarding/components/MinimalistOnboardingForm.tsx
@@ -5,8 +5,8 @@ import Image from 'next/image';
 import { useRouter, useSearchParams } from 'next/navigation';
 import { useCallback, useEffect, useMemo, useState } from 'react';
 import { completeOnboarding } from '@/app/onboarding/actions';
-import { LoadingSpinner } from '@/components/atoms/LoadingSpinner';
 import { Button } from '@/components/ui/Button';
+import { LoadingSpinner } from '@/components/ui/LoadingSpinner';
 import { type HandleValidationState } from '@/components/ui/SmartHandleInput';
 import { APP_URL } from '@/constants/app';
 import { useArtistSearch } from '@/lib/hooks/useArtistSearch';
