--- components/atoms/LoadingSpinner.tsx
@@ -1,1 +1,0 @@
-export { Spinner as LoadingSpinner } from '@/components/ui/Spinner';
