+++ components/ui/PrimaryCTA.tsx
@@ -2,7 +2,7 @@
 
 import { clsx } from 'clsx';
 import React from 'react';
-import { LoadingSpinner } from '@/components/atoms/LoadingSpinner';
+import { LoadingSpinner } from '@/components/ui/LoadingSpinner';
 
 /**
  * PrimaryCTA Component
