+++ components/ui/index.ts
@@ -1,6 +1,5 @@
 // Core UI Components
 
-export { LoadingSpinner } from '@/components/atoms/LoadingSpinner';
 export { ArtistCard } from './ArtistCard';
 export { Badge } from './Badge';
 export { Button } from './Button';
@@ -16,6 +15,7 @@ export { FormStatus } from './FormStatus';
 export { InfoBox } from './InfoBox';
 export { Input } from './Input';
 export { LoadingSkeleton } from './LoadingSkeleton';
+export { LoadingSpinner } from './LoadingSpinner';
 export { Logo } from './Logo';
 export { OptimizedImage } from './OptimizedImage';
 export { PlaceholderImage } from './PlaceholderImage';
@@ -23,6 +23,5 @@ export { ProgressIndicator } from './ProgressIndicator';
 export { Section } from './Section';
 export { Select } from './Select';
 export { SmartHandleInput } from './SmartHandleInput';
-export { Spinner } from './Spinner'; // Legacy - use LoadingSpinner from atoms instead
 export { Textarea } from './Textarea';
 export { VerifiedBadge } from './VerifiedBadge';
